requirements.txt
Flask==1.1.2
matplotlib==3.1.1
numpy==1.18.1
opencv-python-headless==4.2.0.34
gunicorn==20.0.4
tensorflow==2.0.0rc1
scikit-image==0.17.2
firebase==3.0.1
gcloud==0.18.3
gevent==20.6.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud-speech==1.3.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
imageio==2.8.0
imageio-ffmpeg==0.4.2
Jinja2==2.11.2
joblib==0.14.1
keyboard==0.13.5
nltk==3.5
moviepy==1.0.3
librosa==0.7.2
PyAudio==0.2.11
Pyrebase4==4.3.0
pycryptodome==3.9.8